# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ByPass API",
        "description": "API de gestion de bypass réalisée par JOBS Conseil. Cette API permet de gérer les demandes de bypass, les équipements, les zones, les capteurs et les utilisateurs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API ByPass"
        }
    ],
    "paths": {
...
        "/equipment": {
            "get": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Liste des équipements",
                "description": "Récupère la liste des équipements avec filtres optionnels",
                "operationId": "c16a39fcc2c9e03d0e2697d67712f329",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "operational",
                                "maintenance",
                                "down",
                                "standby"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher dans nom ou localisation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des équipements paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Equipment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Créer un équipement",
                "description": "Crée un nouvel équipement. Accessible uniquement aux administrateurs.",
                "operationId": "2e3df0557dce454e87831f194bd9319b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "criticite",
                                    "fabricant",
                                    "zone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Équipement 1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Capteur de température"
                                    },
                                    "criticite": {
                                        "type": "string",
                                        "example": "Haute"
                                    },
                                    "fabricant": {
                                        "type": "string",
                                        "example": "Fabricant XYZ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de l'équipement",
                                        "nullable": true
                                    },
                                    "zone": {
                                        "type": "string",
                                        "example": "Zone A"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "operational",
                                            "maintenance",
                                            "down",
                                            "standby"
                                        ],
                                        "example": "operational"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Équipement créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipment"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "422": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/equipment/{equipment}": {
            "get": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Détails d'un équipement",
                "description": "Récupère les détails d'un équipement spécifique avec ses capteurs",
                "operationId": "f3bb1568b4b0fc98743a8f6a0933e525",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'équipement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipment"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Mettre à jour un équipement",
                "description": "Met à jour les informations d'un équipement. Accessible uniquement aux administrateurs.",
                "operationId": "26a55e30c8daf8776281bdb5cb7c1418",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Équipement 1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Capteur de température"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "operational",
                                            "maintenance",
                                            "down",
                                            "standby"
                                        ],
                                        "example": "operational"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Équipement mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Equipment"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Équipements"
                ],
                "summary": "Supprimer un équipement",
                "description": "Supprime un équipement. Accessible uniquement aux administrateurs.",
                "operationId": "36d5b210acd8759607043f303026cff1",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID de l'équipement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Équipement supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Équipement supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Equipment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "EQ-ZA-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Équipement 1"
                    },
                    "type": {
                        "type": "string",
                        "example": "Capteur de température"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "operational",
                            "maintenance",
                            "down",
                            "standby"
                        ],
                        "example": "operational"
                    },
                    "zone_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token d'authentification Bearer (obtenu via /api/auth/login)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}